Master Mode pets and more

+++ Tiles/Relics/BaseRelic.cs
@@ -91,7 +91,7 @@ namespace SpiritMod.Tiles.Relics
 			// Some math magic to make it smoothly move up and down over time
 			const float TwoPi = (float)Math.PI * 2f;
 			float offset = (float)Math.Sin(Main.GlobalTimeWrappedHourly * TwoPi / 5f);
-			Vector2 drawPos = worldPos + offScreen - Main.screenPosition + new Vector2(0f, -40f) + new Vector2(0f, offset * 4f);
+			Vector2 drawPos = worldPos + offScreen - Main.screenPosition + new Vector2(0f, -40f) + new Vector2(0f, (int)(offset * 4f));
 
 			spriteBatch.Draw(texture, drawPos, frame, color, 0f, origin, 1f, effects, 0f); //Draws the floating relic
